DebugReplay: each 3d fix has a valid gps altitude
Don't try to guess it's availability based on the height value, but only
on the validity flag of the B record.

The IGC file spec says at A2.4 Units: GNSS altitude drop-outs should
be logged as zero altitude. This is obviously only for convenience, as
A4.1 states that 2d fixes (no GPS altitude) should be flagged with a 'V'.
The latter should always have precedence over the zeroed altitude value
and therefore it should be safe to assume a valid gps altitude for
all 3d fixes.