WaypointRenderer: revert "if MAT, draw all Turnpoints as Task points"
Revert commit be927b2b2208a756c5ad0208f2b57deacff6b5b8

This was a bad idea, because it results in a very confusing display
when zoomed out, and the upper limit of 256 visible waypoints is
reached quickly, resulting in artifacts that are even more confusing.
Let's drop this nuisance.