UISettings, ...: remove the "custom fonts" feature
Customizing fonts can break the user interface easily, and is not
portable.  It all relies on the LOGFONT structure, which isn't even
portable well across Windows CE and Windows - let alone Android, Linux
and so on.

Soon, this feature will be replaced with a single knob "scale", which
allows the user to choose larger or smaller fonts.