OS X Retina support
Specify our app to be high resolution capable. Unfortunately, the OS
(and the SDL API) report "points" instead of pixels in this mode. To
determine the correct size of the drawable area, SDL_GL_GetDrawableSize
must be used instead of SDL_GetWindowSize. And input event coordinates
need to be translated to real pixel coordinates.