Computer/Wind: skip WindStore if EKF wind is available
The EKF algorithm is far superior to the circling algorithm.  While
the circling algorithm is unfiltered, EKF is filtered by definition
and does not need the WindStore.  Applying additional filtering is
harmful because it increases the latency.  Let's skip WindStore if EKF
is available.  As a side effect, the circling algorithm gets disabled
completely in that case - which is a good thing.