Airspace/Nearest: don't use Airspaces::FindNearest()
FindNearest() and its backend method KDTree::find_nearest_if() finds
only the nearest bounding box, but for many airspaces, that's a very
bad choice.  It leads to miscalculations (TRAC #3589).  By visiting
all airspaces in range and using AbstractAirspace::ClosestPoint(), we
get the accurate result.  That's much more expensive, but that's the
price we have to pay for accurate calculations.