Markers: manage all markers in Waypoints
Before this commit, we had a big subsystem only for managing markers,
which brought limitations of its own: they could not be used in tasks,
and they could not be used with "goto".  This commit changes all
markers to be just waypoints.  A disadvantage is that markers now lose
they timestamp, because waypoints don't have one - but it remains to
be seen if this is really a big deal.