Android/NativeView: keep existing EGLDisplay and EGLContext
When the EGLSurface gets destroyed (because XCSoar is "paused"), we
don't need to destroy the EGLDisplay and the EGLContext as well.  We
can just reuse them for the next eglCreateWindowSurface() call.

This is the trick that allows us to keep all OpenGL textures while
XCSoar is "paused".  It will render our GLSurfaceListener interface
obsolete.