CrossSection/Airspace: fix missing airspaces in cross section
By default, Airspaces::VisitIntersecting() omits airspaces when the
given vector is completely inside the airspace - which is correct,
because there's no intersection with the airspace outline.  This made
these airspaces invisible in the cross section.

By taking advantage of the new "include_inside" parameter, we can show
these airspaces (TRAC #3537).