Data/Input/default.xci: Remove "Task Restart" menu button
Commit 21b7ef445f0d831be493e8570f016424cda2aa57 introduced a "Task
Restart" menu button. The function activated by this button is only useful
under extremely esoteric conditions, i.e. having a valid start *AND*
achieving at least the next turnpoint *AND* the desire to completely
restart the task from the beginning.

If the button is hit accidentally it will ruin an in-progress task.
Given the limited usefulness for the majority of users it should not
be present by default - those pilots who have some use for it can still
enable it with a custom .xci file.