Screen/Memory/Buffer: constexpr/const fixups for At()
In C++11, "constexpr" on a method implies "const", which is why
WritableImageBuffer::At() has been working at all for const instances
(something we need all the time).  I now learned that clang's
-Wconstexpr-not-const was correct after all.

However this implicit "const" has been removed in C++14, and so this
fails to build.

To make this code future-proof, split WritableImageBuffer::At() into
two overloads: one is "constexpr const", and the other is neither (we
can't have "constexpr" without "const" until we switch to -std=c++14).
After the "-std=c++14" switch (XCSoar 7.0), we can re-add the
"constexpr" to the non-const overload.

The method ConstImageBuffer::At() is also missing the explicit
"const", and will eventually break.