Geo/Flat/Line: correct function return for line with dx<0
This fixes the results of the FlatLine::IntersectOriginCircle() function
for particular cases. The problem manifests itself as incorrect
predicted entry points for a circular airspace area if the predicted
track is westerly – with the error being largest on a track of 225 or
315 degrees.

This fixes the test cases which were added with commit aa7a955.