new InfoBox with "Number Of Satellites"
This (small) modification is to create a new InfoBox with de "Number Of Satellite" used by GPS module.  It can be used to monitor the gps reception quality during the flight, without going to status panel. very usefull for "light users", in paraglider, because we use very light gps modules, sometimes included INTO kobo itself.
The infobox use existing gps status code, and just obtained infos into a conventional InfoBox.  Number of sats and status of connection (3d fix, ....).
As the used words for display are same as originals, the translating system of the app work on it too.
Hope it can be usefull for others.