[Android] Make it possible to play external wav files
This brings SoundUtil functionality on Android on par with WIN32. If the name of the resource to be played ends with .wav it will be loaded as an external file and played. A quick way to test it is with a simple init.lua file:
xcsoar.fire_legacy_event("PlaySound", "/test.wav")