Android: Re-arrange ABI variables, and add documentation
Describe in a comment block where the values of the ABI variables for
Android belong to.
Create a complete block for each architecture, even when values repeat.
Unfortunately no one will completely be matched by another one.
For the sake of consistency and simplicity to create a new set for a new
architecture the values are explicitly written, and transcribed directly
from their respective source.