Doc: French: GPL: 3 --> 2
The French manual contains the text of the license GPL v3, while XCSoar is under GPL v2 (https://github.com/XCSoar/XCSoar/blob/master/COPYING), as mentioned in the English manual.
This PR add the correct version of the GPL license (i.e. V2), with both the English (official) text and an unofficial translation in French.