Driver/KRT2: respond to remote query from radio
Without this, the KRT-2 radio will not report anything to the remote
module. Most important: It will not report the active and passive
frequency.

The main module (i.e. the radio) sends 'S' from time to time. If the
remote module (i.e. XCSoar) answers "with any character within 60 ms",
the main module becomes "active", "indicated by a “r” in the upper right
corner of (the) display".

see "1.1 Activate the Remote connection" and "1.2 Out-going strings" in
the interface manual.